Move ConfirmDeleteModal react component
